+++ test/functional/abc_rpc_excessiveblock.py
@@ -13,9 +13,9 @@ from test_framework.cdefs import (
     ONE_MEGABYTE,
 )
 from test_framework.test_framework import BitcoinTestFramework
-from test_framework.util import assert_equal, assert_raises_rpc_error
+from test_framework.util import assert_equal
 
-BLOCKSIZE_TOO_LOW = "Invalid parameter, excessiveblock must be larger than {}".format(
+BLOCKSIZE_TOO_LOW = "Error: Excessive block size must be > {:,} bytes".format(
     LEGACY_MAX_BLOCK_SIZE)
 
 
@@ -25,7 +25,7 @@ class ExcessiveBlockSizeRPCTest(BitcoinTestFramework):
         self.num_nodes = 1
         self.tip = None
         self.setup_clean_chain = True
-        self.extra_args = [["-deprecatedrpc=setexcessiveblock"]]
+        self.extra_args = [[f"-blockmaxsize={LEGACY_MAX_BLOCK_SIZE}"]]
 
     def check_subversion(self, pattern_str):
         # Check that the subversion is set as expected
@@ -35,50 +35,44 @@ class ExcessiveBlockSizeRPCTest(BitcoinTestFramework):
         assert pattern.match(subversion)
 
     def test_excessiveblock(self):
+        node = self.nodes[0]
+
         # Check that we start with DEFAULT_MAX_BLOCK_SIZE
-        getsize = self.nodes[0].getexcessiveblock()
+        getsize = node.getexcessiveblock()
         ebs = getsize['excessiveBlockSize']
         assert_equal(ebs, DEFAULT_MAX_BLOCK_SIZE)
 
+        def setexcessiveblock(block_size):
+            self.restart_node(
+                0,
+                self.extra_args[0] +
+                [f"-excessiveblocksize={block_size}"])
+
         # Check that setting to legacy size is ok
-        self.nodes[0].setexcessiveblock(LEGACY_MAX_BLOCK_SIZE + 1)
-        getsize = self.nodes[0].getexcessiveblock()
-        ebs = getsize['excessiveBlockSize']
-        assert_equal(ebs, LEGACY_MAX_BLOCK_SIZE + 1)
-
-        # Check that going below legacy size is not accepted
-        assert_raises_rpc_error(-8, BLOCKSIZE_TOO_LOW,
-                                self.nodes[0].setexcessiveblock, LEGACY_MAX_BLOCK_SIZE)
-        getsize = self.nodes[0].getexcessiveblock()
-        ebs = getsize['excessiveBlockSize']
-        assert_equal(ebs, LEGACY_MAX_BLOCK_SIZE + 1)
-
-        # Check that a negative size returns an error
-        assert_raises_rpc_error(-8, BLOCKSIZE_TOO_LOW,
-                                self.nodes[0].setexcessiveblock, -2 * LEGACY_MAX_BLOCK_SIZE)
-        getsize = self.nodes[0].getexcessiveblock()
+        setexcessiveblock(LEGACY_MAX_BLOCK_SIZE + 1)
+        getsize = node.getexcessiveblock()
         ebs = getsize['excessiveBlockSize']
         assert_equal(ebs, LEGACY_MAX_BLOCK_SIZE + 1)
 
         # Check setting to 2MB
-        self.nodes[0].setexcessiveblock(2 * ONE_MEGABYTE)
-        getsize = self.nodes[0].getexcessiveblock()
+        setexcessiveblock(2 * ONE_MEGABYTE)
+        getsize = node.getexcessiveblock()
         ebs = getsize['excessiveBlockSize']
         assert_equal(ebs, 2 * ONE_MEGABYTE)
         # Check for EB correctness in the subver string
         self.check_subversion(r"/Bitcoin ABC:.*\(EB2\.0; .*\)/")
 
         # Check setting to 13MB
-        self.nodes[0].setexcessiveblock(13 * ONE_MEGABYTE)
-        getsize = self.nodes[0].getexcessiveblock()
+        setexcessiveblock(13 * ONE_MEGABYTE)
+        getsize = node.getexcessiveblock()
         ebs = getsize['excessiveBlockSize']
         assert_equal(ebs, 13 * ONE_MEGABYTE)
         # Check for EB correctness in the subver string
         self.check_subversion(r"/Bitcoin ABC:.*\(EB13\.0; .*\)/")
 
         # Check setting to 13.14MB
-        self.nodes[0].setexcessiveblock(13140000)
-        getsize = self.nodes[0].getexcessiveblock()
+        setexcessiveblock(13140000)
+        getsize = node.getexcessiveblock()
         ebs = getsize['excessiveBlockSize']
         assert_equal(ebs, 13.14 * ONE_MEGABYTE)
         # check for EB correctness in the subver string